requirements.txt
Django==2.2.6
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-social-share==1.4.0
django-social-widgets==0.4.0
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
